Completed 'Register' API route and frontend UI
This was a fairly large commit, but overall, these changes led to the
completion of the '/register' endpoint and frontend user-interface.

Created the SQLite database and have it hooked up to the '/register'
route. The request body is validated using Joi and when successful,
creates a new user based off 'models/User'.

The frontend which utilizes the '/register' route to allow users a way
to sign up uses Vue and Bulma, with heavy reliance on Buefy.