AddSite.vue - adds the ability to add site via UI
The newest component allows users to add a site via the user-interface.
The UI for adding a new site is a form inside a Buefy modal which is
opened by clicking the '+ Add Site' button (only available for logged
in users). When pressed, the modal opens up, regardless of the current
view (because its parent component is the Header.vue, shown on all pag-
es) and users can input the necessary information for adding a fire
site to the app.