Refactoring and start of ViewSite.vue
ViewSite.vue will be the component used to display information for a s-
ingle site (when the url is: '.../sites/example-site'). The request pa-
sses a single parameter (siteId) and a successful response replies with
the Site model matching the parameter from the databse. This site data
is then used to create a detailed page for users to interact with (such
as favoriting the site, seeing similar sites based on the tags given,
etc.).