Fixed err 'no-side-effects-in-computed-properties'
The computed method `sortSites` was throwing said error. Being a newb
to Vue and its practices, I didn't realize that it is considered bad
practice to introduce side effects (aka directly altering data) inside
computed properties. So instead of sorting the `sites` data array dire-
ctly, simply slicing the array (making a copy) and modifying that is
a better approach.

Refs: https://github.com/vuejs/eslint-plugin-vue/blob/master/docs/rules/no-side-effects-in-computed-properties.md