Update Registration to include a `name` field
This commit adds a `name` column to the User model and adds the necess-
ary validationi steps to make sure a string is typed. Also included in
is the frontend updates that adds an additional input field to the Reg-
ister.vue component.

These changes are made so user information is not needed to be inputted
in the AddSite.vue component (specifically the `created_by` and eventu-
ally the `user_location` values) and can just be pulled from the user
object stored using Vuex.