Update Site cards && implement search
This is the beginning of a larger frontend refactoring / refresh. The
UI will eventually have a more modern look and feel with the use of more
vibrant colors, shadows, and animations.

Also, search functionality was implemented and was meant to have its own
commit, but I messed up somewhere and now its included in this commit :shrug: