Move from Bulma columns to tiles - fixes #7
On wider layouts, the site cards are laid in three columns. On mobile,
these columns collapse into a single column, and the order needs to
be preserved. Previously, there were three computed functions that
broke up the sites in the three Bulma columns for wider layouts, but that
did not work for mobile. For example:

     **** screen > 768px ****
        _ _ _ _ _ _ _ _ _ _
        |  0  |  1  |  2  |
        - - - - - - - - - -
        |  3  |  4  |  5  |
        - - - - - - - - - -
        |  6  |  7  |  8  |
        - - - - - - - - - -

      **** screen < 768px ****
              _ _ _ _
              |  0  |
              - - - -
              |  3  |
              - - - -
              |  6  |
              - - - -
              |  1  |
              - - - -
              | ... |
              - - - -

Notice the order of the mobile layout. To correct this, the use of
Bulma tiles was implemented, which allows for creating *rows* which
will spill in proper order on smaller screen sizes. Tiles use a
nesting hierarchy of at least three layers: ancestor (the row),
parent (the column), and the child (actual content). Because we
are dynamically fetching our data from the server, hard-coding these
elements won't work. Instead, some math is involved to create the
correct number of rows and properly fill each row's columns (just 3
wide).

To determine the correct number of rows (`<div.tile.is-ancestor`),
we take the length of the `sites` array, divide it by 3 (no. columns)
rounded to the nearest integer greater or equal to the quotient. That
gives us the number of rows:

    `v-for="row in Math.ceil(sites.length / 3)"`

Next, to dynamically give each row three columns, we need to slice the
`sites` array into groups of threes based on the row number. Row 1,
for example, should contain elements from the `sites` array with
indices 0 through 2; Row 2 will have elements 3 through 5, and so
forth. The math for figuring out the starting and ending index to slice
the `sites` array for each row is as follows.

    Begin index: (x * 3) - 3        where x = row number (begining w/ 1)
    End index: (x * 2) + (x [+ 1])  remember `Arr.slice()` is exclusive
                                              so the `+ 1` to the row number is dropped

This math appears in the following line of code:

    `v-for="site in sites.slice((row * 3) - 3, (row * 2) + row)"`

All of this makes it so that when the three columns of Site cards are
collapsed on narrower layouts, the order in which they appear remains intact.