Always display 'Add Site' + 'Profile' buttons
This commit fixes #16 and completes the second task from #12. Reworked
the "Profile" dropdown menu to either show 'Sign In / Out' if the user is
not logged in, or the previous account-specific links ('Favorites', 'Acc-
ount', etc.) when logged in. By removing the sign in / out buttons from the
navbar, we can always display the "Add Site" button and have it redirect to
the login page if the user is not logged in.

The fix for #16 was to remove the `hoverable` property from the `b-dropdown`
element, which causes a modal with the menu items to appear on smaller screen
sizes.