Fix issue with related site displaying favorites
Assume Site A has a favorite. Clicking on Site A from the 'sites' route
will take you the detailed page for Site A. Below, Sites X, Y, and Z are
shown as 'related sites'. Since Site A has a favorite, the star icon is
filled in.

The problem: when clicking on any site from the 'related sites' section,
the `favorite` data key does not get updated with the new site. So if
Site A was favorited, Site X was not and Site X was clicked, the favorited
star will be shown as filled in for Site X.

The solution: because the initial getting of favorites occurs when the comp-
onent `mounts`, the function to get a sites function was not being triggered
when reusing the `ViewSite.vue` component (i.e. clicking on a related site).
To make this fetch happen, we move the code that obtains a site's favorite
from the `mounted` function to the `methods` section where all of the regetting
of site data when navigating between 'related sites' occurs.